refatorado implementação de brokers, atualziado para modelo usando functions
