--- ds-conta-corrente-produtos/src/main/java/com/farias/banco/dscontacorrenteprodutos/feignclients/PageMixInCustom.java
@@ -1,8 +1,0 @@
-package com.farias.banco.dscontacorrenteprodutos.feignclients;
-
-import com.farias.banco.dscontacorrenteprodutos.feignclients.impl.PageResourcesImpl;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-
-
-@JsonDeserialize(as = PageResourcesImpl.class)
-public interface PageMixInCustom{ }
