--- ds-conta-corrente-produtos/src/main/java/com/farias/banco/dscontacorrenteprodutos/feignclients/impl/CustomJacksonModule.java
@@ -1,16 +1,0 @@
-package com.farias.banco.dscontacorrenteprodutos.feignclients.impl;
-
-import org.springframework.data.domain.Page;
-
-import com.farias.banco.dscontacorrenteprodutos.feignclients.PageMixInCustom;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-
-public class CustomJacksonModule extends SimpleModule {
-
-	private static final long serialVersionUID = 1L;
-
-	@Override
-    public void setupModule(SetupContext context) {
-        context.setMixInAnnotations(Page.class, PageMixInCustom.class);
-    }
-}
