+++ ds-conta-corrente-produtos/src/main/java/com/farias/banco/dscontacorrenteprodutos/resources/ContaCorrenteProdutosResource.java
@@ -1,24 +1,22 @@
  package com.farias.banco.dscontacorrenteprodutos.resources;
 
-import java.util.List;
-import java.util.Optional;
+ import com.farias.banco.dscontacorrenteprodutos.dto.ContaCorrenteProdutoDTO;
+ import com.farias.banco.dscontacorrenteprodutos.service.ContaCorrenteProdutosService;
+ import io.swagger.annotations.Api;
+ import io.swagger.annotations.ApiOperation;
+ import io.swagger.annotations.ApiResponse;
+ import io.swagger.annotations.ApiResponses;
+ import lombok.RequiredArgsConstructor;
+ import org.springframework.data.domain.Page;
+ import org.springframework.data.domain.PageRequest;
+ import org.springframework.data.domain.Sort;
+ import org.springframework.web.bind.annotation.GetMapping;
+ import org.springframework.web.bind.annotation.RequestMapping;
+ import org.springframework.web.bind.annotation.RequestParam;
+ import org.springframework.web.bind.annotation.RestController;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.farias.banco.dscontacorrenteprodutos.dto.ContaCorrenteProdutoDTO;
-import com.farias.banco.dscontacorrenteprodutos.service.ContaCorrenteProdutosService;
-
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import lombok.RequiredArgsConstructor;
+ import java.util.List;
+ import java.util.Optional;
 
 @Api(tags = {"Conta corrente produtos"})
 @RequiredArgsConstructor
